Sound and Music into own classes and got sound FX playback in remaster.
