+++ Core/AV/Sound/WAVEFORMATEX.cs
@@ -1,0 +1,23 @@
+﻿using System.Runtime.InteropServices;
+
+namespace OpenVIII.AV
+{
+    public static partial class Sound
+    {
+        #region Structs
+
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Ansi, Pack = 1)]
+        public struct WAVEFORMATEX
+        {
+            public ushort wFormatTag;
+            public ushort nChannels;
+            public uint nSamplesPerSec;
+            public uint nAvgBytesPerSec;
+            public ushort nBlockAlign;
+            public ushort wBitsPerSample;
+            public ushort cbSize;
+        }
+
+        #endregion Structs
+    }
+}
