+++ Core/Core.csproj
@@ -147,18 +147,26 @@
   <ItemGroup>
     <Compile Include="Archive\Archive.cs" />
     <Compile Include="Archive\ArchiveBase.cs" />
+    <Compile Include="Archive\ArchiveMap.cs" />
     <Compile Include="Archive\Archives.cs" />
     <Compile Include="Archive\ArchiveWorker.cs" />
+    <Compile Include="Archive\FI.cs" />
+    <Compile Include="Archive\FL.cs" />
     <Compile Include="Archive\ZZZ\ArchiveZZZ.cs" />
     <Compile Include="Archive\ZZZ\FileData.cs" />
     <Compile Include="Archive\ZZZ\Header.cs" />
     <Compile Include="AV\Audio.cs" />
     <Compile Include="AV\Movie\Directories.cs" />
     <Compile Include="AV\Movie\Files.cs" />
-    <Compile Include="AV\Midi\Fluid.cs" />
-    <Compile Include="AV\Midi\DirectMedia.cs" />
+    <Compile Include="AV\Music\Midi\Fluid.cs" />
+    <Compile Include="AV\Music\Midi\DirectMedia.cs" />
     <Compile Include="AV\Movie\Player.cs" />
     <Compile Include="AV\Movie\STATE.cs" />
+    <Compile Include="AV\Music\Music.cs" />
+    <Compile Include="AV\Sound\ADPCMCOEFSET.cs" />
+    <Compile Include="AV\Sound\Entry.cs" />
+    <Compile Include="AV\Sound\Sound.cs" />
+    <Compile Include="AV\Sound\WAVEFORMATEX.cs" />
     <Compile Include="AV\Video.cs" />
     <Compile Include="Battle\AnimationSystem.cs" />
     <Compile Include="Battle\BattleSwirl.cs" />
@@ -212,7 +220,7 @@
     <Compile Include="Battle\Stage\Model.cs" />
     <Compile Include="Battle\Stage\ModelGroup.cs" />
     <Compile Include="Battle\RegularPyramid.cs" />
-    <Compile Include="AV\Buffer_Data.cs" />
+    <Compile Include="AV\Sound\Buffer_Data.cs" />
     <Compile Include="Battle\Stage\ObjectsGroup.cs" />
     <Compile Include="Battle\Stage\Polygon.cs" />
     <Compile Include="Battle\Stage\Quad.cs" />
@@ -839,7 +847,6 @@
     <Compile Include="Font.cs" />
     <Compile Include="Game1.cs" />
     <Compile Include="GFs.cs" />
-    <Compile Include="AV\init_debugger_Audio.cs" />
     <Compile Include="Kernel\Kernel_bin.Abilities.cs" />
     <Compile Include="Kernel\Kernel_bin.Ability.cs" />
     <Compile Include="Kernel\Kernel_bin.AbilityFlags2.cs" />
@@ -955,7 +962,7 @@
     <Compile Include="Strings\Strings.Areames.cs" />
     <Compile Include="Strings\Strings.cs" />
     <Compile Include="ModuleHandler.cs" />
-    <Compile Include="AV\module_movie_test.cs" />
+    <Compile Include="AV\Movie\module_movie_test.cs" />
     <Compile Include="module_overture_debug.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Saves\Saves.CharacterData.cs" />
