+++ Core/Field/Module.cs
@@ -125,7 +125,7 @@ namespace OpenVIII.Fields
             // lets you move through all the feilds just holding left or right. it will just loop when it runs out.
             //if (false && Input2.DelayedButton(FF8TextTagKey.Left))
             //{
-            //    init_debugger_Audio.PlaySound(0);
+            //    AV.Sound.Play(0);
             //    if (Memory.FieldHolder.FieldID > 0)
             //        Memory.FieldHolder.FieldID--;
             //    else
@@ -134,7 +134,7 @@ namespace OpenVIII.Fields
             //}
             //else if (false && Input2.DelayedButton(FF8TextTagKey.Right))
             //{
-            //    init_debugger_Audio.PlaySound(0);
+            //    AV.Sound.Play(0);
             //    if (Memory.FieldHolder.FieldID < checked((ushort)(Memory.FieldHolder.fields.Length - 1)))
             //        Memory.FieldHolder.FieldID++;
             //    else
