+++ Core/World/module_world_debug.cs
@@ -268,7 +268,7 @@ namespace OpenVIII
             }
             //temporarily disabling this, because I'm getting more and more tired of this music playing over and over when debugging
             //Memory.musicIndex = 30;
-            //init_debugger_Audio.PlayMusic();
+            //AV.Music.Play();
             ate = new AlphaTestEffect(Memory.graphics.GraphicsDevice)
             {
                 Projection = projectionMatrix,
